Tests/LibWeb/Text/expected/wheel-event-should-be-consumed-by-innermost-scrollable.txt

inner.scrollTop: 100, inner.scrollLeft: 0
